# Improve stack usage of all the Handle functions