@cdvr1993 2023-07-21
> Are there any benchmarks you could run?

Added the benchmark code. The result is here:

### No assert
#### Previous
<img width="824" alt="Screenshot 2023-07-21 at 3 26 17 PM" src="https://github.com/yarpc/yarpc-go/assets/3793727/e88ce72e-ea64-4802-8b20-886f6dab97ca">

#### Now
<img width="824" alt="Screenshot 2023-07-21 at 3 27 16 PM" src="https://github.com/yarpc/yarpc-go/assets/3793727/49a2305d-14a3-4702-9f6d-668cab363b95">


### With assert
#### Previous
<img width="803" alt="Screenshot 2023-07-21 at 3 29 59 PM" src="https://github.com/yarpc/yarpc-go/assets/3793727/4f80670d-b1df-4126-a065-8c43841b25be">

#### Now
<img width="821" alt="Screenshot 2023-07-21 at 3 29 49 PM" src="https://github.com/yarpc/yarpc-go/assets/3793727/0ab2c16d-f470-4b2c-85dd-52274d8d8b42">