@DheerendraRathor 2023-07-22
@cdvr1993 I ran benchmarks with and without graph changes (keeping just call & zap changes) on my laptop. 

Without graph changes:
```
$ go test -bench=BenchmarkMiddlewareHandle -benchmem
goos: darwin
goarch: amd64
pkg: go.uber.org/yarpc/internal/observability
cpu: Intel(R) Core(TM) i7-1068NG7 CPU @ 2.30GHz
BenchmarkMiddlewareHandle-8      2292955               517.4 ns/op           192 B/op          2 allocs/op
PASS
ok      go.uber.org/yarpc/internal/observability        2.240s
```

With graph changes (this PR as-is)
```
$ go test -bench=BenchmarkMiddlewareHandle -benchmem
goos: darwin
goarch: amd64
pkg: go.uber.org/yarpc/internal/observability
cpu: Intel(R) Core(TM) i7-1068NG7 CPU @ 2.30GHz
BenchmarkMiddlewareHandle-8      2327730               515.1 ns/op           192 B/op          2 allocs/op
PASS
ok      go.uber.org/yarpc/internal/observability        2.289s

```

There isn't a significant difference in both.