@cdvr1993 2023-07-24
@DheerendraRathor Ok, I updated it. It seems the benchmark is faster as you said. Although stack usage is greater:

```
middleware.go:183     0x62d6f                 4881ec60020000          SUBQ $0x260, SP
```

vs

```
  middleware.go:183     0x5c728                 4881ec50010000                  SUBQ $0x150, SP
```

If we want to have more common and readable code removing only fields from call is the way to do it.

If we want to reduce as much as possible stack usage in the middleware function, the original PR is the best. I'll let you to decide.