@DheerendraRathor 2023-07-25
@cdvr1993 I think there is some miscommunication here between comments. 
I've no objection in converting all call methods to pointer receiver unless there is huge gain in performance with value receivers which I don't think is the case. This will also help you in stack use reduction. 

I had objection in changes to `graph.begin` API which IMO reduced the code readability and maintainability.