@jronak 2023-07-25
I can't seem to look at the history of the PR, but what is the difference in the benchmarking numbers between the current changes vs with pointer receivers? 

Thanks for this @cdvr1993!