@cdvr1993 2023-07-26
@jronak 0x988 vs 0x260, so 1.8KB