@cdvr1993 2023-07-26
@rabbbit last benchmark results
```
go test ./internal/observability/... -bench . -count=10
goos: linux
goarch: amd64
pkg: go.uber.org/yarpc/internal/observability
cpu: AMD EPYC 7B13
BenchmarkMiddlewareHandle-96             1429688               844.9 ns/op
BenchmarkMiddlewareHandle-96             1443886               828.4 ns/op
BenchmarkMiddlewareHandle-96             1442847               828.4 ns/op
BenchmarkMiddlewareHandle-96             1451000               821.1 ns/op
BenchmarkMiddlewareHandle-96             1442511               833.8 ns/op
BenchmarkMiddlewareHandle-96             1436607               834.3 ns/op
BenchmarkMiddlewareHandle-96             1462096               815.1 ns/op
BenchmarkMiddlewareHandle-96             1425666               816.2 ns/op
BenchmarkMiddlewareHandle-96             1450243               824.1 ns/op
BenchmarkMiddlewareHandle-96             1458702               815.5 ns/op
PASS
ok      go.uber.org/yarpc/internal/observability        20.577s
```