@rabbbit 2023-07-26
Do you know why the allocations were reduced too?